running lint

+++ python/rapidsmp/rapidsmp/shuffler.pxd
@@ -1,6 +1,6 @@
 # Copyright (c) 2025, NVIDIA CORPORATION.
 
-from libc.stdint cimport uint32_t, uint16_t
+from libc.stdint cimport uint16_t, uint32_t
 from libcpp cimport bool
 from libcpp.memory cimport shared_ptr, unique_ptr
 from libcpp.string cimport string
